Moved containerService to its own file.
